+++ packages/shared/CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 Please note, that this project, while following numbering syntax, it DOES NOT
 adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) rules.
 
+## [5.19.1] - 2026-06-17
+
+### Fixed
+- `ChartSerializer` will no longer duplicate the GeoJSON data for user-defined items.
+- `MapChart` `projectionName` was not resolving in standalone script/CDN bundles. The bundled projections are now registered whenever `am5map` is loaded.
+
+
 ## [5.19.0] - 2026-06-15
 
 ### Added